refactor(sc2): split galaxy_host.c sc2_* natives into per-subject headers
14 implementation-include headers created in games/starcraft-2/game/galaxy/:
  galaxy_math.h, galaxy_trigger.h, galaxy_camera.h, galaxy_cinematic.h,
  galaxy_unit.h, galaxy_point.h, galaxy_player.h, galaxy_sound.h,
  galaxy_transmission.h, galaxy_catalog.h, galaxy_game.h, galaxy_ui.h,
  galaxy_techtree.h, galaxy_actor.h

galaxy_host.c shrinks from ~1600 to ~770 lines (lifecycle + state + dispatch
table only). Each domain can now be extended without touching the host file.
