+++ games/starcraft-2/game/galaxy/galaxy_actor.h
@@ -1,0 +1,36 @@
+/* galaxy_actor.h — achievement, actor, bank, conversation, and timer natives */
+static DWORD sc2_AchievementAward(LPJASS j)              { (void)j; return jass_pushnull(j); }
+static DWORD sc2_AchievementErase(LPJASS j)              { (void)j; return jass_pushnull(j); }
+static DWORD sc2_AchievementPanelSetCategory(LPJASS j)   { (void)j; return jass_pushnull(j); }
+static DWORD sc2_AchievementPanelSetVisible(LPJASS j)    { (void)j; return jass_pushnull(j); }
+static DWORD sc2_AchievementPercentText(LPJASS j)        { (void)j; return jass_pushnull(j); }
+static DWORD sc2_AchievementTermQuantitySet(LPJASS j)    { (void)j; return jass_pushnull(j); }
+static DWORD sc2_AchievementsDisable(LPJASS j)           { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ActorCreate(LPJASS j)                   { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ActorFrom(LPJASS j)                     { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ActorFromScope(LPJASS j)                { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ActorRegionCreate(LPJASS j)             { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ActorRegionSend(LPJASS j)               { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ActorScopeFromUnit(LPJASS j)            { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ActorSend(LPJASS j)                     { (void)j; return jass_pushnull(j); }
+static DWORD sc2_BankExists(LPJASS j)                    { (void)j; return jass_pushboolean(j, false); }
+static DWORD sc2_BankKeyRemove(LPJASS j)                 { (void)j; return jass_pushnull(j); }
+static DWORD sc2_BankLastCreated(LPJASS j)               { (void)j; return jass_pushnullhandle(j, "bank"); }
+static DWORD sc2_BankLoad(LPJASS j)                      { (void)j; return jass_pushnullhandle(j, "bank"); }
+static DWORD sc2_BankSave(LPJASS j)                      { (void)j; return jass_pushnull(j); }
+static DWORD sc2_BankValueSetFromFlag(LPJASS j)          { (void)j; return jass_pushnull(j); }
+static DWORD sc2_BankValueSetFromInt(LPJASS j)           { (void)j; return jass_pushnull(j); }
+static DWORD sc2_BankValueSetFromString(LPJASS j)        { (void)j; return jass_pushnull(j); }
+static DWORD sc2_BankValueSetFromText(LPJASS j)          { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ConversationDataResetNodeState(LPJASS j)  { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ConversationDataResetStateValues(LPJASS j){ (void)j; return jass_pushnull(j); }
+static DWORD sc2_ConversationDataSaveNodeState(LPJASS j)   { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ConversationDataSaveStateValues(LPJASS j) { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ConversationDataStateFixedValue(LPJASS j) { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ConversationDataStateGetValue(LPJASS j)   { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ConversationDataStateIndex(LPJASS j)      { (void)j; return jass_pushinteger(j, 0); }
+static DWORD sc2_ConversationDataStateIndexCount(LPJASS j) { (void)j; return jass_pushinteger(j, 0); }
+static DWORD sc2_ConversationDataStateName(LPJASS j)       { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ConversationDataStateSetValue(LPJASS j)   { (void)j; return jass_pushnull(j); }
+static DWORD sc2_ConversationDataStateText(LPJASS j)       { (void)j; return jass_pushnull(j); }
+static DWORD sc2_TimerPause(LPJASS j)                    { (void)j; return jass_pushnull(j); }
