+++ games/starcraft-2/game/galaxy/galaxy_host.c
@@ -333,858 +333,20 @@ void galaxy_tick(LPJASS vm) {
     }
 }
 
-/* -------------------------------------------------------------------------
- * Trigger natives
- * ------------------------------------------------------------------------- */
-/* Triggers are opaque handles; extract the integer ID via pointer cast. */
-static LONG sc2_trigger_id(LPJASS j, int index) {
-    HANDLE h = jass_checkhandle(j, index, "trigger");
-    return h ? (LONG)(uintptr_t)h : 0;
-}
-
-static DWORD sc2_TriggerCreate(LPJASS j) {
-    LPCSTR name = jass_checkstring(j, 1);
-    LONG id = 0;
-    if (name && sc2_trig_n < MAX_SC2_TRIGGERS) {
-        id = sc2_trig_next_id++;
-        /* strdup: the JASS VM may GC the string after this call returns. */
-        sc2_trigs[sc2_trig_n++] = (sc2trig_t){ id, strdup(name), false };
-    }
-    return id ? jass_pushlighthandle(j, (HANDLE)(uintptr_t)id, "trigger")
-              : jass_pushnullhandle(j, "trigger");
-}
-
-static DWORD sc2_TriggerEnable(LPJASS j)      { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerStop(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerGetCurrent(LPJASS j)  { return jass_pushnullhandle(j, "trigger"); }
-static DWORD sc2_TriggerGetExecCount(LPJASS j){ return jass_pushinteger(j, 0); }
-static DWORD sc2_TriggerIsEnabled(LPJASS j)   { return jass_pushboolean(j, true); }
-
-static DWORD sc2_TriggerExecute(LPJASS j) {
-    LONG   id        = sc2_trigger_id(j, 1);
-    BOOL   testConds = jass_checkboolean(j, 2);
-    BOOL   waitDone  = jass_checkboolean(j, 3);
-    for (DWORD i = 0; i < sc2_trig_n; i++) {
-        if (sc2_trigs[i].id == id && sc2_trigs[i].func) {
-#ifdef SC2_DEBUG_CUTSCENE
-            fprintf(stderr, "TriggerExecute: %s testConds=%d waitDone=%d\n",
-                    sc2_trigs[i].func, testConds, waitDone);
-#endif
-            sc2_fire_trigger_func(j, sc2_trigs[i].func, testConds, !waitDone);
-            break;
-        }
-    }
-    return jass_pushnull(j);
-}
-
-static DWORD sc2_TriggerAddEventMapInit(LPJASS j) {
-    LONG id = sc2_trigger_id(j, 1);
-    for (DWORD i = 0; i < sc2_trig_n; i++) {
-        if (sc2_trigs[i].id == id) { sc2_trigs[i].mapinit = true; break; }
-    }
-    return jass_pushnull(j);
-}
-
-static DWORD sc2_TriggerSkippableBegin(LPJASS j) { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerSkippableEnd(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerQueueEnter(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerQueueExit(LPJASS j)      { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerQueueIsEmpty(LPJASS j)   { return jass_pushboolean(j, true); }
-static DWORD sc2_TriggerQueuePause(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerQueueClear(LPJASS j)     { (void)j; return jass_pushnull(j); }
-
-static DWORD sc2_Wait(LPJASS j) {
-    FLOAT secs = jass_checknumber(j, 1);
-    jass_sleep(j, (DWORD)(secs * 1000.0f));
-    return 0;
-}
-
-/* Event registration stubs */
-static DWORD sc2_TriggerAddEventPlayerAIWave(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventPlayerAllianceChange(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventPlayerLeft(LPJASS j)              { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventPlayerPropChange(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventTimeElapsed(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventTimePeriodic(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventTimer(LPJASS j)                   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventUnitAttacked(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventUnitCargo(LPJASS j)               { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventUnitDamaged(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventUnitDied(LPJASS j)                { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventUnitOrder(LPJASS j)               { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventUnitRange(LPJASS j)               { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventUnitRangePoint(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerAddEventUnitRegion(LPJASS j)              { (void)j; return jass_pushnull(j); }
-
-/* -------------------------------------------------------------------------
- * Camera
- * ------------------------------------------------------------------------- */
-
-/* CameraInfoFromId: look up map camera by ID, store in local table, return handle. */
-static DWORD sc2_CameraInfoFromId(LPJASS j) {
-    DWORD map_id = (DWORD)jass_checkinteger(j, 1);
-    FLOAT tx = 0, ty = 0, tz = 0, pitch = 56.0f, yaw = 180.0f, dist = 34.0f, fov = 28.0f, height = 0;
-    if (sc2_galaxy_get_camera_by_id &&
-        sc2_gcam_n < MAX_GALAXY_CAMS &&
-        sc2_galaxy_get_camera_by_id(map_id, &tx, &ty, &tz, &pitch, &yaw, &dist, &fov, &height)) {
-        LONG h = sc2_gcam_n++;
-        sc2_gcams[h] = (sc2GCam_t){ tx, ty, tz, pitch, yaw, dist, fov, height };
-        return jass_pushlighthandle(j, (HANDLE)(uintptr_t)h, "camerainfo");
-    }
-    return jass_pushnullhandle(j, "camerainfo");
-}
-
-static DWORD sc2_CameraInfoDefault(LPJASS j) { return jass_pushnullhandle(j, "camerainfo"); }
-
-/* CameraApplyInfo: apply stored camera to client (duration = 0 → instant snap). */
-static DWORD sc2_CameraApplyInfo(LPJASS j) {
-    LONG h = (LONG)(uintptr_t)jass_checkhandle(j, 2, "camerainfo");
-    FLOAT dur = jass_checknumber(j, 3);
-#ifdef SC2_DEBUG_CUTSCENE
-    fprintf(stderr, "CameraApplyInfo: handle=%ld count=%ld duration=%.2f callback=%d\n",
-            (long)h, (long)sc2_gcam_n, dur, sc2_galaxy_on_camera != NULL);
-#endif
-    if (h > 0 && h < sc2_gcam_n && sc2_galaxy_on_camera) {
-        sc2GCam_t *c = &sc2_gcams[h];
-        sc2_galaxy_on_camera(c->tx, c->ty, c->yaw, c->pitch, c->dist, c->fov, c->height, dur);
-    }
-    return jass_pushnull(j);
-}
-
-static DWORD sc2_CameraPan(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_CameraSave(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_CameraRestore(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_CameraLockInput(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_CameraShakeStart(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_CameraGetTarget(LPJASS j)   { return jass_pushinteger(j, 0); }
-
-/* -------------------------------------------------------------------------
- * Cinematic
- * ------------------------------------------------------------------------- */
-static DWORD sc2_CinematicMode(LPJASS j) {
-    BOOL  enable = jass_checkboolean(j, 2);
-    FLOAT dur    = jass_checknumber(j, 3);
-#ifdef SC2_DEBUG_CUTSCENE
-    fprintf(stderr, "CinematicMode: enable=%d dur=%.1f\n", enable, dur);
-#endif
-    if (sc2_galaxy_on_cinematic) sc2_galaxy_on_cinematic(enable, dur);
-    return jass_pushnull(j);
-}
-
-static DWORD sc2_CinematicFade(LPJASS j) {
-    BOOL  fadein = jass_checkboolean(j, 1);
-    FLOAT dur    = jass_checknumber(j, 2);
-#ifdef SC2_DEBUG_CUTSCENE
-    fprintf(stderr, "CinematicFade: fadein=%d dur=%.1f\n", fadein, dur);
-#endif
-    if (sc2_galaxy_on_fade) sc2_galaxy_on_fade(fadein ? 0.0f : 1.0f, dur);
-    return jass_pushnull(j);
-}
-
-static DWORD sc2_CinematicOverlay(LPJASS j) { (void)j; return jass_pushnull(j); }
-
-/* -------------------------------------------------------------------------
- * Units
- * UnitCreate: resolve unit model from catalog, spawn at point position.
- * ------------------------------------------------------------------------- */
-static DWORD sc2_UnitCreate(LPJASS j) {
-    LONG   count  = jass_checkinteger(j, 1);
-    LPCSTR type   = jass_checkstring(j, 2);
-    LONG   player = jass_checkinteger(j, 4);
-    LONG   pt_h   = (LONG)(uintptr_t)jass_checkhandle(j, 5, "point");
-    FLOAT  angle  = jass_checknumber(j, 6);
-    FLOAT  x = 0.0f, y = 0.0f;
-    if (pt_h > 0 && pt_h < sc2_gpoint_n) { x = sc2_gpoints[pt_h].x; y = sc2_gpoints[pt_h].y; }
-    LONG handle = 0;
-    for (LONG i = 0; i < count && sc2_gunit_n < MAX_GALAXY_UNITS; i++) {
-        void *ent = sc2_galaxy_on_unit_create ?
-            sc2_galaxy_on_unit_create(type ? type : "", (int)player, x, y, angle) : NULL;
-        if (!ent)
-            fprintf(stderr, "sc2_UnitCreate: on_unit_create returned NULL for type '%s' (%ld/%ld) — unit will be invisible\n",
-                    type ? type : "(null)", (long)(i + 1), (long)count);
-        handle = (LONG)(++sc2_gunit_n);
-        sc2_gunits[handle - 1] = ent;
-        sc2_last_unit_handle = handle;
-    }
-    return handle ? jass_pushlighthandle(j, (HANDLE)(uintptr_t)handle, "unit")
-                  : jass_pushnullhandle(j, "unit");
-}
-
-static DWORD sc2_UnitLastCreated(LPJASS j) {
-    return sc2_last_unit_handle ?
-        jass_pushlighthandle(j, (HANDLE)(uintptr_t)sc2_last_unit_handle, "unit") :
-        jass_pushnullhandle(j, "unit");
-}
-static DWORD sc2_UnitLastCreatedGroup(LPJASS j)     { return jass_pushnullhandle(j, "unitgroup"); }
-static void *sc2_ent_from_handle(LPJASS j, int idx) {
-    LONG h = (LONG)(uintptr_t)jass_checkhandle(j, idx, "unit");
-    return (h > 0 && h <= (LONG)sc2_gunit_n) ? sc2_gunits[h - 1] : NULL;
-}
-
-static DWORD sc2_UnitSetPosition(LPJASS j) {
-    void *ent = sc2_ent_from_handle(j, 1);
-    LONG pt_h = (LONG)(uintptr_t)jass_checkhandle(j, 2, "point");
-    if (ent && sc2_galaxy_unit_set_position && pt_h > 0 && pt_h < sc2_gpoint_n)
-        sc2_galaxy_unit_set_position(ent, sc2_gpoints[pt_h].x, sc2_gpoints[pt_h].y, 0.0f);
-    return jass_pushnull(j);
-}
-
-static DWORD sc2_UnitSetFacing(LPJASS j) {
-    void *ent = sc2_ent_from_handle(j, 1);
-    FLOAT ang = jass_checknumber(j, 2);
-    if (ent && sc2_galaxy_unit_set_position) {
-        /* Re-use set_position with NaN for x/y to indicate facing-only update.
-         * g_sc2.c checks for this sentinel and only updates the angle. */
-        sc2_galaxy_unit_set_position(ent, 0.0f/0.0f, 0.0f/0.0f, ang * 3.14159265f / 180.0f);
-    }
-    return jass_pushnull(j);
-}
-
-static DWORD sc2_UnitSetOwner(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitGetOwner(LPJASS j) {
-    (void)jass_checkhandle(j, 1, "unit");  /* consume arg */
-    return jass_pushinteger(j, 0);
-}
-static DWORD sc2_UnitSetHeight(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitSetScale(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitSetState(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitSetCursor(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitTestState(LPJASS j)            { return jass_pushboolean(j, false); }
-static DWORD sc2_UnitIsAlive(LPJASS j) {
-    void *ent = sc2_ent_from_handle(j, 1);
-    return jass_pushboolean(j, ent && (!sc2_galaxy_unit_is_alive || sc2_galaxy_unit_is_alive(ent)));
-}
-static DWORD sc2_UnitIsValid(LPJASS j) {
-    void *ent = sc2_ent_from_handle(j, 1);
-    return jass_pushboolean(j, ent != NULL);
-}
-static DWORD sc2_UnitKill(LPJASS j)                 { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitRemove(LPJASS j)               { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitRevive(LPJASS j)               { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitWaitUntilIdle(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitIssueOrder(LPJASS j) {
-    LONG unit_h  = (LONG)(uintptr_t)jass_checkhandle(j, 1, "unit");
-    LONG order_h = (LONG)(uintptr_t)jass_checkhandle(j, 2, "order");
-    LONG queue   = jass_checkinteger(j, 3);
-    if (order_h <= 0 || order_h >= sc2_gorder_n)
-        return jass_pushboolean(j, false);
-    sc2GOrder_t *ord = &sc2_gorders[order_h];
-    LONG  ac_h = ord->abilcmd_h;
-    LONG  pt_h = ord->pt_h;
-    FLOAT tx   = (pt_h > 0 && pt_h < sc2_gpoint_n) ? sc2_gpoints[pt_h].x : 0.0f;
-    FLOAT ty   = (pt_h > 0 && pt_h < sc2_gpoint_n) ? sc2_gpoints[pt_h].y : 0.0f;
-    const char *ability = (ac_h > 0 && ac_h < sc2_gabilcmd_n)
-                          ? sc2_gabilcmds[ac_h].ability : "move";
-    if (unit_h <= 0 || unit_h > (LONG)sc2_gunit_n || (queue != 0 && queue != 1))
-        return jass_pushboolean(j, false);
-    if (queue == 0) sc2_uorder_n[unit_h - 1] = 0;
-    LONG *count = &sc2_uorder_n[unit_h - 1];
-    if (*count >= MAX_UNIT_ORDERS) {
-        fprintf(stderr, "UnitIssueOrder: queue full for unit %ld\n", (long)unit_h);
-        return jass_pushboolean(j, false);
-    }
-    sc2GUnitOrder_t *queued = &sc2_uorders[unit_h - 1][(*count)++];
-    snprintf(queued->ability, sizeof(queued->ability), "%s", ability);
-    queued->x = tx;
-    queued->y = ty;
-    queued->started = false;
-#ifdef SC2_DEBUG_CUTSCENE
-    fprintf(stderr, "UnitIssueOrder: unit=%ld ability=%s target=(%.1f,%.1f) queue=%ld depth=%ld\n",
-            (long)unit_h, ability, tx, ty, (long)queue, (long)*count);
-#endif
-    return jass_pushboolean(j, true);
-}
-static DWORD sc2_UnitPauseAll(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitGetFacing(LPJASS j)            { return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_UnitGetHeight(LPJASS j)            { return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_UnitGetPosition(LPJASS j) {
-    return jass_pushnullhandle(j, "point");  /* TODO: extract ent position */
-}
-static DWORD sc2_UnitGetType(LPJASS j)   { (void)jass_checkhandle(j, 1, "unit"); return jass_pushstring(j, ""); }
-static DWORD sc2_UnitFromId(LPJASS j)    { return jass_pushnullhandle(j, "unit"); }
-static DWORD sc2_UnitBehaviorAdd(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitBehaviorRemove(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitCargoCreate(LPJASS j) {
-    LONG   t_h  = (LONG)(uintptr_t)jass_checkhandle(j, 1, "unit");
-    LPCSTR type = jass_checkstring(j, 2);
-    LONG   cnt  = jass_checkinteger(j, 3);
-    if (cnt < 1) cnt = 1;
-    LONG handle  = 0;
-    for (LONG i = 0; i < cnt && sc2_gunit_n < MAX_GALAXY_UNITS; i++) {
-        void *ent = sc2_galaxy_on_unit_create ?
-            sc2_galaxy_on_unit_create(type ? type : "", 0, 0.0f, 0.0f, 0.0f) : NULL;
-        if (!ent)
-            fprintf(stderr, "sc2_UnitCargoCreate: on_unit_create returned NULL for type '%s' (%ld/%ld) — cargo unit will be invisible\n",
-                    type ? type : "(null)", (long)(i + 1), (long)cnt);
-        handle = (LONG)(++sc2_gunit_n);
-        sc2_gunits[handle - 1] = ent;
-        sc2_last_cargo_handle  = handle;
-        sc2_last_unit_handle   = handle;
-        if (t_h > 0 && t_h <= MAX_GALAXY_UNITS) {
-            LONG ci = sc2_gcargo_n[t_h - 1];
-            if (ci < MAX_CARGO_PER_UNIT)
-                sc2_gcargo[t_h - 1][sc2_gcargo_n[t_h - 1]++] = handle;
-        }
-    }
-    fprintf(stderr, "UnitCargoCreate: transport=%ld type=%s count=%ld\n",
-            (long)t_h, type ? type : "(null)", (long)cnt);
-    return handle ? jass_pushlighthandle(j, (HANDLE)(uintptr_t)handle, "unit")
-                  : jass_pushnullhandle(j, "unit");
-}
-static DWORD sc2_UnitCargoLastCreated(LPJASS j) {
-    return sc2_last_cargo_handle ?
-        jass_pushlighthandle(j, (HANDLE)(uintptr_t)sc2_last_cargo_handle, "unit") :
-        jass_pushnullhandle(j, "unit");
-}
-static DWORD sc2_UnitCargoGroup(LPJASS j) {
-    LONG h = (LONG)(uintptr_t)jass_checkhandle(j, 1, "unit");
-    if (h > 0 && h <= (LONG)sc2_gunit_n)
-        return jass_pushlighthandle(j,
-            (HANDLE)(uintptr_t)(CARGO_GROUP_FLAG | h), "unitgroup");
-    return jass_pushnullhandle(j, "unitgroup");
-}
-static DWORD sc2_UnitCargoLastCreatedGroup(LPJASS j){ return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitClearSelection(LPJASS j)       { (void)j; return jass_pushnull(j); }
-/* UnitRef wraps a unit handle into a unitref (same pointer, different type name). */
-static DWORD sc2_UnitRefFromUnit(LPJASS j) {
-    HANDLE h = jass_checkhandle(j, 1, "unit");
-    return h ? jass_pushlighthandle(j, h, "unitref") : jass_pushnullhandle(j, "unitref");
-}
-static DWORD sc2_UnitRefFromVariable(LPJASS j)  { return jass_pushnullhandle(j, "unitref"); }
-static DWORD sc2_UnitRefToUnit(LPJASS j) {
-    HANDLE h = jass_checkhandle(j, 1, "unitref");
-    return h ? jass_pushlighthandle(j, h, "unit") : jass_pushnullhandle(j, "unit");
-}
-static DWORD sc2_UnitSetInfoText(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitClearInfoText(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitForceStatusBar(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitGetAttachmentPoint(LPJASS j)   { return jass_pushinteger(j, 0); }
-static DWORD sc2_UnitSetTeamColorIndex(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitLoadModel(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitUnloadModel(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitSetPropertyFixed(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitGetPropertyFixed(LPJASS j)     { return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_EventUnit(LPJASS j)       { return jass_pushnullhandle(j, "unit"); }
-static DWORD sc2_EventUnitCargo(LPJASS j)  { return jass_pushnullhandle(j, "unit"); }
-static DWORD sc2_EventUnitTarget(LPJASS j) { return jass_pushnullhandle(j, "unit"); }
-
-/* UnitGroup */
-static DWORD sc2_UnitGroupEmpty(LPJASS j)            { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupAdd(LPJASS j)              { return jass_checkhandle(j, 1, "unitgroup") ? jass_pushlighthandle(j, jass_checkhandle(j, 1, "unitgroup"), "unitgroup") : jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupCount(LPJASS j) {
-    LONG h    = (LONG)(uintptr_t)jass_checkhandle(j, 1, "unitgroup");
-    LONG mode = jass_checkinteger(j, 2); (void)mode;
-    if (h & CARGO_GROUP_FLAG) {
-        LONG t = h & ~CARGO_GROUP_FLAG;
-        if (t > 0 && t <= (LONG)sc2_gunit_n)
-            return jass_pushinteger(j, sc2_gcargo_n[t - 1]);
-    }
-    return jass_pushinteger(j, 0);
-}
-static DWORD sc2_UnitGroupHasUnit(LPJASS j)          { return jass_pushboolean(j, false); }
-static DWORD sc2_UnitGroupWaitUntilIdle(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitInventoryGroup(LPJASS j)        { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitTechTreeBehaviorCount(LPJASS j) { return jass_pushinteger(j, 0); }
-static DWORD sc2_UnitTechTreeUnitCount(LPJASS j)     { return jass_pushinteger(j, 0); }
-static DWORD sc2_UnitTechTreeUpgradeCount(LPJASS j)  { return jass_pushinteger(j, 0); }
-static DWORD sc2_UnitGroupUnit(LPJASS j)             { return jass_pushinteger(j, 0); }
-static DWORD sc2_UnitGroupIssueOrder(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitGroupIdle(LPJASS j)             { return jass_pushboolean(j, false); }
-static DWORD sc2_UnitGroupFilter(LPJASS j)           { return jass_pushinteger(j, 0); }
-static DWORD sc2_UnitGroupClear(LPJASS j)            { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupCopy(LPJASS j)             { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupRemove(LPJASS j)           { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitGroupAlliance(LPJASS j)         { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupFilterAlliance(LPJASS j)   { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupFilterPlayer(LPJASS j)     { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupFilterPlane(LPJASS j)      { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupFilterRegion(LPJASS j)     { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupFilterThreat(LPJASS j)     { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupFromId(LPJASS j)           { return jass_pushnullhandle(j, "unitgroup"); }
-static DWORD sc2_UnitGroupLoopBegin(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitGroupLoopCurrent(LPJASS j)      { return jass_pushnullhandle(j, "unit"); }
-static DWORD sc2_UnitGroupLoopDone(LPJASS j)         { return jass_pushboolean(j, true); }
-static DWORD sc2_UnitGroupLoopEnd(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitGroupLoopStep(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitGroupNearestUnit(LPJASS j)      { return jass_pushnullhandle(j, "unit"); }
-static DWORD sc2_UnitGroupRandomUnit(LPJASS j)       { return jass_pushnullhandle(j, "unit"); }
-static DWORD sc2_UnitGroupTestPlane(LPJASS j)        { return jass_pushboolean(j, false); }
-static DWORD sc2_UnitFilter(LPJASS j)                { return jass_pushnullhandle(j, "unitfilter"); }
-static DWORD sc2_UnitFilterMatch(LPJASS j)           { return jass_pushboolean(j, false); }
-static DWORD sc2_UnitFilterSetState(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitFilterStr(LPJASS j)             { return jass_pushnullhandle(j, "unitfilter"); }
-static DWORD sc2_UnitGroup(LPJASS j)                 { return jass_pushnullhandle(j, "unitgroup"); }
-
-/* UnitType */
-static DWORD sc2_UnitTypeFromString(LPJASS j)          { return jass_pushstring(j, ""); }
-static DWORD sc2_UnitTypeGetCost(LPJASS j)             { return jass_pushinteger(j, 0); }
-static DWORD sc2_UnitTypeGetName(LPJASS j)             { return jass_pushstring(j, ""); }
-static DWORD sc2_UnitTypeGetProperty(LPJASS j)         { return jass_pushinteger(j, 0); }
-static DWORD sc2_UnitTypeIsAffectedByUpgrade(LPJASS j) { return jass_pushboolean(j, false); }
-static DWORD sc2_UnitTypeTestAttribute(LPJASS j)       { return jass_pushboolean(j, false); }
-static DWORD sc2_UnitTypeTestFlag(LPJASS j)            { return jass_pushboolean(j, false); }
-static DWORD sc2_UnitTypeAnimationLoad(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitTypeAnimationUnload(LPJASS j)     { (void)j; return jass_pushnull(j); }
-
-/* -------------------------------------------------------------------------
- * Point, Region, Visibility
- * ------------------------------------------------------------------------- */
-/* PointFromId: look up map point object by ID, return typed handle. */
-static DWORD sc2_PointFromId(LPJASS j) {
-    DWORD map_id = (DWORD)jass_checkinteger(j, 1);
-    FLOAT x = 0.0f, y = 0.0f;
-    if (sc2_galaxy_get_point_by_id && sc2_gpoint_n < MAX_GALAXY_POINTS &&
-        sc2_galaxy_get_point_by_id(map_id, &x, &y)) {
-        LONG h = sc2_gpoint_n++;
-        sc2_gpoints[h] = (sc2GPoint_t){ x, y };
-        return jass_pushlighthandle(j, (HANDLE)(uintptr_t)h, "point");
-    }
-    if (sc2_gpoint_n >= MAX_GALAXY_POINTS)
-        fprintf(stderr, "PointFromId: table full (%d entries) — id=%u lost\n",
-                MAX_GALAXY_POINTS, map_id);
-    return jass_pushnullhandle(j, "point");
-}
-
-/* Point(x, y): create a point from explicit coordinates. */
-static DWORD sc2_Point(LPJASS j) {
-    FLOAT x = jass_checknumber(j, 1), y = jass_checknumber(j, 2);
-    if (sc2_gpoint_n < MAX_GALAXY_POINTS) {
-        LONG h = sc2_gpoint_n++;
-        sc2_gpoints[h] = (sc2GPoint_t){ x, y };
-        return jass_pushlighthandle(j, (HANDLE)(uintptr_t)h, "point");
-    }
-    return jass_pushnullhandle(j, "point");
-}
-
-static FLOAT sc2_point_x(LPJASS j, int idx) {
-    LONG h = (LONG)(uintptr_t)jass_checkhandle(j, idx, "point");
-    return (h > 0 && h < sc2_gpoint_n) ? sc2_gpoints[h].x : 0.0f;
-}
-static FLOAT sc2_point_y(LPJASS j, int idx) {
-    LONG h = (LONG)(uintptr_t)jass_checkhandle(j, idx, "point");
-    return (h > 0 && h < sc2_gpoint_n) ? sc2_gpoints[h].y : 0.0f;
-}
-
-static DWORD sc2_PointGetX(LPJASS j) { return jass_pushnumber(j, sc2_point_x(j, 1)); }
-static DWORD sc2_PointGetY(LPJASS j) { return jass_pushnumber(j, sc2_point_y(j, 1)); }
-static DWORD sc2_PointGetHeight(LPJASS j)        { return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_PointGetFacing(LPJASS j)        { return jass_pushnumber(j, 0.0f); }
-
-/* PointWithOffset: returns a new point offset from the base. */
-static DWORD sc2_PointWithOffset(LPJASS j) {
-    FLOAT bx = sc2_point_x(j, 1), by = sc2_point_y(j, 1);
-    FLOAT dx = jass_checknumber(j, 2), dy = jass_checknumber(j, 3);
-    if (sc2_gpoint_n < MAX_GALAXY_POINTS) {
-        LONG h = sc2_gpoint_n++;
-        sc2_gpoints[h] = (sc2GPoint_t){ bx + dx, by + dy };
-        return jass_pushlighthandle(j, (HANDLE)(uintptr_t)h, "point");
-    }
-    return jass_pushnullhandle(j, "point");
-}
-
-/* PointWithOffsetPolar: returns a new point offset by (dist, angle_deg). */
-static DWORD sc2_PointWithOffsetPolar(LPJASS j) {
-    FLOAT bx = sc2_point_x(j, 1), by = sc2_point_y(j, 1);
-    FLOAT dist = jass_checknumber(j, 2);
-    FLOAT ang  = jass_checknumber(j, 3) * 3.14159265f / 180.0f; /* degrees → radians */
-    if (sc2_gpoint_n < MAX_GALAXY_POINTS) {
-        LONG h = sc2_gpoint_n++;
-        sc2_gpoints[h] = (sc2GPoint_t){ bx + dist * cosf(ang), by + dist * sinf(ang) };
-        return jass_pushlighthandle(j, (HANDLE)(uintptr_t)h, "point");
-    }
-    return jass_pushnullhandle(j, "point");
-}
-
-static DWORD sc2_PointReflect(LPJASS j)          { return jass_pushnullhandle(j, "point"); }
-static DWORD sc2_PointPathingCliffLevel(LPJASS j){ return jass_pushinteger(j, 0); }
-
-static DWORD sc2_AngleBetweenPoints(LPJASS j) {
-    FLOAT ax = sc2_point_x(j, 1), ay = sc2_point_y(j, 1);
-    FLOAT bx = sc2_point_x(j, 2), by = sc2_point_y(j, 2);
-    return jass_pushnumber(j, atan2f(by - ay, bx - ax) * 180.0f / 3.14159265f);
-}
-
-static DWORD sc2_DistanceBetweenPoints(LPJASS j) {
-    FLOAT ax = sc2_point_x(j, 1), ay = sc2_point_y(j, 1);
-    FLOAT bx = sc2_point_x(j, 2), by = sc2_point_y(j, 2);
-    FLOAT dx = bx - ax, dy = by - ay;
-    return jass_pushnumber(j, sqrtf(dx*dx + dy*dy));
-}
-
-static DWORD sc2_RegionEmpty(LPJASS j)           { return jass_pushnullhandle(j, "region"); }
-static DWORD sc2_RegionEntireMap(LPJASS j)       { return jass_pushnullhandle(j, "region"); }
-static DWORD sc2_RegionPlayableMap(LPJASS j)     { return jass_pushnullhandle(j, "region"); }
-static DWORD sc2_RegionPlayableMapSet(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_RegionRect(LPJASS j)            { return jass_pushnullhandle(j, "region"); }
-static DWORD sc2_RegionCircle(LPJASS j)          { return jass_pushnullhandle(j, "region"); }
-static DWORD sc2_RegionAddCircle(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_RegionAddRect(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_RegionAddRegion(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_RegionAttachToUnit(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_RegionContainsPoint(LPJASS j)   { return jass_pushboolean(j, false); }
-static DWORD sc2_RegionFromId(LPJASS j)          { return jass_pushnullhandle(j, "region"); }
-static DWORD sc2_RegionGetAttachUnit(LPJASS j)   { return jass_pushnullhandle(j, "unit"); }
-static DWORD sc2_RegionGetBoundsMax(LPJASS j)    { return jass_pushnullhandle(j, "point"); }
-static DWORD sc2_RegionGetBoundsMin(LPJASS j)    { return jass_pushnullhandle(j, "point"); }
-static DWORD sc2_RegionGetCenter(LPJASS j)       { return jass_pushnullhandle(j, "point"); }
-static DWORD sc2_RegionGetOffset(LPJASS j)       { return jass_pushnullhandle(j, "point"); }
-static DWORD sc2_RegionRandomPoint(LPJASS j)     { return jass_pushnullhandle(j, "point"); }
-static DWORD sc2_RegionSetCenter(LPJASS j)       { (void)j; return jass_pushnull(j); }
-
-static DWORD sc2_VisRevealArea(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_VisExploreArea(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_VisRevealerCreate(LPJASS j)     { return jass_pushnullhandle(j, "revealer"); }
-static DWORD sc2_VisRevealerLastCreated(LPJASS j){ return jass_pushnullhandle(j, "revealer"); }
-static DWORD sc2_VisRevealerDestroy(LPJASS j)    { (void)j; return jass_pushnull(j); }
-
-/* -------------------------------------------------------------------------
- * Player / PlayerGroup
- * ------------------------------------------------------------------------- */
-static DWORD sc2_PlayerGroupAll(LPJASS j)            { return jass_pushnullhandle(j, "playergroup"); }
-static DWORD sc2_PlayerGroupActive(LPJASS j)         { return jass_pushnullhandle(j, "playergroup"); }
-static DWORD sc2_PlayerGroupAdd(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerGroupClear(LPJASS j)          { return jass_pushnullhandle(j, "playergroup"); }
-static DWORD sc2_PlayerGroupCopy(LPJASS j)           { return jass_pushnullhandle(j, "playergroup"); }
-/* PlayerGroupCount / PlayerGroupPlayer: singleplayer stubs — exactly 1 human player. */
-static DWORD sc2_PlayerGroupCount(LPJASS j) {
-    (void)jass_checkhandle(j, 1, "playergroup");
-    return jass_pushinteger(j, 1);
-}
-static DWORD sc2_PlayerGroupEmpty(LPJASS j)          { return jass_pushnullhandle(j, "playergroup"); }
-static DWORD sc2_PlayerGroupHasPlayer(LPJASS j)      { return jass_pushboolean(j, false); }
-static DWORD sc2_PlayerGroupPlayer(LPJASS j) {
-    (void)jass_checkhandle(j, 1, "playergroup");
-    (void)jass_checkinteger(j, 2);
-    return jass_pushinteger(j, 1);
-}
-static DWORD sc2_PlayerGroupLoopBegin(LPJASS j)      { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerGroupLoopDone(LPJASS j)       { return jass_pushboolean(j, true); }
-static DWORD sc2_PlayerGroupLoopEnd(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerGroupLoopStep(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerGroupRemove(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerGroupSingle(LPJASS j)         { return jass_pushnullhandle(j, "playergroup"); }
-static DWORD sc2_PlayerGroupAlliance(LPJASS j)       { return jass_pushnullhandle(j, "playergroup"); }
-static DWORD sc2_PlayerGetState(LPJASS j)            { return jass_pushinteger(j, 0); }
-static DWORD sc2_PlayerSetState(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerGetAlliance(LPJASS j)         { return jass_pushboolean(j, false); }
-static DWORD sc2_PlayerSetAlliance(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerDifficulty(LPJASS j)          { return jass_pushinteger(j, 0); }
-static DWORD sc2_PlayerModifyPropertyInt(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerAddChargeRegen(LPJASS j)      { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerAddChargeUsed(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerAddCooldown(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerGetChargeRegen(LPJASS j)      { return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_PlayerGetChargeUsed(LPJASS j)       { return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_PlayerGetCooldown(LPJASS j)         { return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_PlayerScoreValueEnable(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerScoreValueEnableAll(LPJASS j) { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerScoreValueGetAsFixed(LPJASS j){ return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_PlayerScoreValueGetAsInt(LPJASS j)  { return jass_pushinteger(j, 0); }
-static DWORD sc2_PlayerScoreValueSetFromFixed(LPJASS j){ (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerScoreValueSetFromInt(LPJASS j){ (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerPauseAllCharges(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerPauseAllCooldowns(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerBeaconAlert(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerBeaconClearTarget(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerBeaconGetTargetPoint(LPJASS j){ return jass_pushinteger(j, 0); }
-static DWORD sc2_PlayerBeaconGetTargetUnit(LPJASS j) { return jass_pushinteger(j, 0); }
-static DWORD sc2_PlayerBeaconIsAutoCast(LPJASS j)    { return jass_pushboolean(j, false); }
-static DWORD sc2_PlayerBeaconIsFromUser(LPJASS j)    { return jass_pushboolean(j, false); }
-static DWORD sc2_PlayerBeaconIsSet(LPJASS j)         { return jass_pushboolean(j, false); }
-static DWORD sc2_PlayerBeaconSetAutoCast(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerBeaconSetTargetPoint(LPJASS j){ (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerBeaconSetTargetUnit(LPJASS j) { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerCreateEffectPoint(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerCreateEffectUnit(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerValidateEffectPoint(LPJASS j) { return jass_pushboolean(j, false); }
-static DWORD sc2_PlayerValidateEffectUnit(LPJASS j)  { return jass_pushboolean(j, false); }
-
-/* -------------------------------------------------------------------------
- * Sound
- * ------------------------------------------------------------------------- */
-static DWORD sc2_SoundLink(LPJASS j) {
-    LPCSTR id = jass_checkstring(j, 1);
-    LONG asset = jass_checkinteger(j, 2), h;
-    if (!id || !*id || sc2_gsound_n >= MAX_GALAXY_SOUNDS)
-        return jass_pushnullhandle(j, "soundlink");
-    h = sc2_gsound_n++;
-    snprintf(sc2_gsounds[h].id, sizeof(sc2_gsounds[h].id), "%s", id);
-    sc2_gsounds[h].asset = asset;
-    return jass_pushlighthandle(j, (HANDLE)(uintptr_t)h, "soundlink");
-}
-static DWORD sc2_SoundLinkAsset(LPJASS j)     { return jass_pushnullhandle(j, "soundlink"); }
-static DWORD sc2_SoundLinkId(LPJASS j)        { return jass_pushnullhandle(j, "soundlink"); }
-static DWORD sc2_SoundPlay(LPJASS j)          { return jass_pushnullhandle(j, "sound"); }
-static DWORD sc2_SoundPlayAtPoint(LPJASS j)   { return jass_pushnullhandle(j, "sound"); }
-static DWORD sc2_SoundPlayOnUnit(LPJASS j)    { return jass_pushnullhandle(j, "sound"); }
-static DWORD sc2_SoundPlayScene(LPJASS j)     { return jass_pushnullhandle(j, "sound"); }
-static DWORD sc2_SoundPlaySceneFile(LPJASS j) { return jass_pushnullhandle(j, "sound"); }
-static DWORD sc2_SoundStop(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_SoundWait(LPJASS j) {
-    FLOAT secs = jass_checknumber(j, 2);
-    if (secs > 0.0f) jass_sleep(j, (DWORD)(secs * 1000.0f));
-    return jass_pushnull(j);
-}
-static FLOAT sc2_sound_length(LPJASS j, int arg) {
-    LONG h = (LONG)(uintptr_t)jass_checkhandle(j, arg, "soundlink");
-    return h > 0 && h < sc2_gsound_n && sc2_galaxy_sound_length
-        ? sc2_galaxy_sound_length(sc2_gsounds[h].id, sc2_gsounds[h].asset) : 0.0f;
-}
-static DWORD sc2_SoundLengthSync(LPJASS j)    { return jass_pushnumber(j, sc2_sound_length(j, 1)); }
-static DWORD sc2_SoundtrackPlay(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_SoundtrackPause(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_SoundtrackDefault(LPJASS j)  { (void)j; return jass_pushnull(j); }
-
-/* -------------------------------------------------------------------------
- * Math
- * ------------------------------------------------------------------------- */
-static DWORD sc2_AbsF(LPJASS j)  { FLOAT x = jass_checknumber(j, 1); return jass_pushnumber(j, x < 0 ? -x : x); }
-static DWORD sc2_MaxF(LPJASS j)  { FLOAT a = jass_checknumber(j,1), b = jass_checknumber(j,2); return jass_pushnumber(j, a > b ? a : b); }
-static DWORD sc2_MinF(LPJASS j)  { FLOAT a = jass_checknumber(j,1), b = jass_checknumber(j,2); return jass_pushnumber(j, a < b ? a : b); }
-static DWORD sc2_ModF(LPJASS j)  { FLOAT a = jass_checknumber(j,1), b = jass_checknumber(j,2); return jass_pushnumber(j, b ? fmodf(a, b) : 0.0f); }
-static DWORD sc2_Pow(LPJASS j)   { return jass_pushnumber(j, powf(jass_checknumber(j,1), jass_checknumber(j,2))); }
-static DWORD sc2_SquareRoot(LPJASS j) { FLOAT x = jass_checknumber(j,1); return jass_pushnumber(j, x > 0.0f ? sqrtf(x) : 0.0f); }
-static DWORD sc2_Sin(LPJASS j)   { return jass_pushnumber(j, sinf(jass_checknumber(j,1))); }
-static DWORD sc2_Cos(LPJASS j)   { return jass_pushnumber(j, cosf(jass_checknumber(j,1))); }
-static DWORD sc2_Tan(LPJASS j)   { return jass_pushnumber(j, tanf(jass_checknumber(j,1))); }
-static DWORD sc2_ASin(LPJASS j)  { return jass_pushnumber(j, asinf(jass_checknumber(j,1))); }
-static DWORD sc2_ACos(LPJASS j)  { return jass_pushnumber(j, acosf(jass_checknumber(j,1))); }
-static DWORD sc2_ATan(LPJASS j)  { return jass_pushnumber(j, atanf(jass_checknumber(j,1))); }
-static DWORD sc2_ATan2(LPJASS j) { return jass_pushnumber(j, atan2f(jass_checknumber(j,1), jass_checknumber(j,2))); }
-static DWORD sc2_RandomFixed(LPJASS j) { (void)j; return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_RandomInt(LPJASS j)   { return jass_pushinteger(j, jass_checkinteger(j, 1)); }
-
-/* -------------------------------------------------------------------------
- * String / Text / Color / Game / AI / Order / Catalog / Objectives / etc.
- * ------------------------------------------------------------------------- */
-static DWORD sc2_StringExternal(LPJASS j)     { LPCSTR s = jass_checkstring(j,1); return jass_pushstring(j, s ? s : ""); }
-/* Galaxy uses a null result to terminate whitespace-delimited word iteration. */
-static DWORD sc2_StringWord(LPJASS j) {
-    LPCSTR str = jass_checkstring(j, 1), word;
-    LONG index = jass_checkinteger(j, 2);
-    if (!str || index < 1) return jass_pushnull(j);
-    while (*str) {
-        while (*str && isspace((unsigned char)*str)) str++;
-        if (!*str) break;
-        word = str;
-        while (*str && !isspace((unsigned char)*str)) str++;
-        if (!--index) return jass_pushstringlen(j, word, (DWORD)(str - word));
-    }
-    return jass_pushnull(j);
-}
-static DWORD sc2_IntToText(LPJASS j)          { return jass_pushinteger(j, 0); }
-static DWORD sc2_Color(LPJASS j)              { (void)j; return jass_pushinteger(j, 0); }
-static DWORD sc2_ColorWithAlpha(LPJASS j)     { (void)j; return jass_pushinteger(j, 0); }
-static DWORD sc2_GameTimeOfDayPause(LPJASS j) { (void)j; return jass_pushnull(j); }
-static DWORD sc2_GameTimeOfDaySet(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_GameSetBackground(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_GameSetLighting(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_DifficultyEnabled(LPJASS j)  { return jass_pushboolean(j, false); }
-static DWORD sc2_DifficultyName(LPJASS j)     { return jass_pushstring(j, ""); }
-static DWORD sc2_DifficultyNameCampaign(LPJASS j) { return jass_pushstring(j, ""); }
-static DWORD sc2_AITimePause(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_AbilityClass(LPJASS j)       { return jass_pushinteger(j, 0); }
-static DWORD sc2_AbilityCommand(LPJASS j) {
-    LPCSTR name = jass_checkstring(j, 1);
-    LONG   cmd  = jass_checkinteger(j, 2);
-    if (sc2_gabilcmd_n < MAX_GALAXY_ABILCMDS) {
-        LONG h = sc2_gabilcmd_n++;
-        snprintf(sc2_gabilcmds[h].ability, sizeof(sc2_gabilcmds[h].ability),
-                 "%s", name ? name : "");
-        sc2_gabilcmds[h].cmd_idx = cmd;
-        return jass_pushinteger(j, h);
-    }
-    fprintf(stderr, "AbilityCommand: table full (%d entries) — '%s' lost\n",
-            MAX_GALAXY_ABILCMDS, name ? name : "");
-    return jass_pushinteger(j, 0);
-}
-static DWORD sc2_AbilityCommandGetAbility(LPJASS j){ return jass_pushstring(j, ""); }
-static DWORD sc2_AbilityCommandGetCommand(LPJASS j){ return jass_pushinteger(j, 0); }
-static DWORD sc2_AbilityCommandGetAction(LPJASS j) { return jass_pushinteger(j, 0); }
-static DWORD sc2_Order(LPJASS j)              { return jass_pushnullhandle(j, "order"); }
-static DWORD sc2_OrderTargetingPoint(LPJASS j) {
-    LONG abilcmd_h = jass_checkinteger(j, 1);
-    LONG pt_h      = (LONG)(uintptr_t)jass_checkhandle(j, 2, "point");
-    if (sc2_gorder_n < MAX_GALAXY_ORDERS) {
-        LONG h = sc2_gorder_n++;
-        sc2_gorders[h] = (sc2GOrder_t){ abilcmd_h, pt_h };
-        return jass_pushlighthandle(j, (HANDLE)(uintptr_t)h, "order");
-    }
-    fprintf(stderr, "OrderTargetingPoint: table full (%d entries)\n", MAX_GALAXY_ORDERS);
-    return jass_pushnullhandle(j, "order");
-}
-static DWORD sc2_OrderTargetingUnit(LPJASS j) { return jass_pushnullhandle(j, "order"); }
-static DWORD sc2_OrderSetPlayer(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UnitOrderIsValid(LPJASS j)   { return jass_pushboolean(j, false); }
-static DWORD sc2_CatalogEntryClass(LPJASS j)  { return jass_pushinteger(j, 0); }
-static DWORD sc2_CatalogEntryCount(LPJASS j)  { return jass_pushinteger(j, 0); }
-static DWORD sc2_CatalogEntryGet(LPJASS j)    { return jass_pushstring(j, ""); }
-static DWORD sc2_CatalogEntryIsValid(LPJASS j){ return jass_pushboolean(j, false); }
-static DWORD sc2_CatalogEntryParent(LPJASS j) { return jass_pushstring(j, ""); }
-static DWORD sc2_CatalogEntryScope(LPJASS j)  { return jass_pushstring(j, ""); }
-static DWORD sc2_CatalogFieldCount(LPJASS j)  { return jass_pushinteger(j, 0); }
-static DWORD sc2_CatalogFieldGet(LPJASS j)    { return jass_pushstring(j, ""); }
-static DWORD sc2_CatalogFieldIsArray(LPJASS j){ return jass_pushboolean(j, false); }
-static DWORD sc2_CatalogFieldIsScope(LPJASS j){ return jass_pushboolean(j, false); }
-static DWORD sc2_CatalogFieldType(LPJASS j)   { return jass_pushstring(j, ""); }
-static DWORD sc2_CatalogFieldValueCount(LPJASS j){ return jass_pushinteger(j, 0); }
-static DWORD sc2_CatalogFieldValueGet(LPJASS j){ return jass_pushstring(j, ""); }
-static DWORD sc2_CatalogFieldValueSet(LPJASS j){ return jass_pushboolean(j, false); }
-static DWORD sc2_ObjectiveCreate3(LPJASS j)   { return jass_pushinteger(j, 0); }
-static DWORD sc2_ObjectiveLastCreated(LPJASS j){ return jass_pushinteger(j, 0); }
-static DWORD sc2_ObjectiveSetName(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ObjectiveSetState(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ObjectiveGetState(LPJASS j)  { return jass_pushinteger(j, 0); }
-static DWORD sc2_PingLastCreated(LPJASS j)    { return jass_pushinteger(j, 0); }
-static DWORD sc2_PingDestroy(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PingSetScale(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PingSetTooltip(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_MinimapPing(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UISetMode(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UIAlertPoint(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UIAlertUnit(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UISetFrameVisible(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_HelpPanelAddTip(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_HelpPanelDisplayPage(LPJASS j){ (void)j; return jass_pushnull(j); }
-static DWORD sc2_HelpPanelEnableTechTreeButton(LPJASS j){ (void)j; return jass_pushnull(j); }
-static DWORD sc2_TransmissionSource(LPJASS j)        { return jass_pushnullhandle(j, "transmissionsource"); }
-static DWORD sc2_TransmissionSourceFromModel(LPJASS j){ return jass_pushnullhandle(j, "transmissionsource"); }
-static DWORD sc2_TransmissionSourceFromUnit(LPJASS j) { return jass_pushnullhandle(j, "transmissionsource"); }
-/* TransmissionSend: send a transmission to players; sleep if waitUntilDone and duration > 0. */
-static DWORD sc2_TransmissionSend(LPJASS j) {
-    /* args: playergroup, source, camerainfo, string anim, soundlink, text speaker,
-     *       text msg, fixed duration, int durationType, bool waitUntilDone */
-    FLOAT sound_dur  = sc2_sound_length(j, 5);
-    FLOAT dur        = jass_checknumber(j, 8);
-    LONG  dur_type   = jass_checkinteger(j, 9);
-    BOOL  wait_done  = jass_checkboolean(j, 10);
-    LONG  sound_h    = (LONG)(uintptr_t)jass_checkhandle(j, 5, "soundlink");
-    /* Native SC2 derives transmission time from the linked asset before applying the requested modifier. */
-    if (dur_type == 0) dur = sound_dur;
-    else if (dur_type == 1) dur += sound_dur;
-    else if (dur_type == 2) dur = MAX(sound_dur - dur, 0.0f);
-    if (sound_h > 0 && sound_h < sc2_gsound_n && sc2_galaxy_on_sound)
-        sc2_galaxy_on_sound(sc2_gsounds[sound_h].id, sc2_gsounds[sound_h].asset);
-    fprintf(stderr, "TransmissionSend: sound=%s asset=%ld duration=%.2f wait=%d\n",
-            sound_h > 0 && sound_h < sc2_gsound_n ? sc2_gsounds[sound_h].id : "(null)",
-            sound_h > 0 && sound_h < sc2_gsound_n ? (long)sc2_gsounds[sound_h].asset : -1L, dur, wait_done);
-    if (wait_done && dur > 0.0f)
-        jass_sleep(j, (DWORD)(dur * 1000.0f));
-    return jass_pushnullhandle(j, "sound");
-}
-static DWORD sc2_TransmissionLastSent(LPJASS j)      { return jass_pushinteger(j, 0); }
-static DWORD sc2_TransmissionClear(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TransmissionClearAll(LPJASS j)      { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TransmissionWait(LPJASS j) {
-    FLOAT secs = jass_checknumber(j, 2);
-    if (secs > 0.0f) jass_sleep(j, (DWORD)(secs * 1000.0f));
-    return jass_pushnull(j);
-}
-static DWORD sc2_TechTreeUpgradeAddLevel(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_VictoryPanelAddAchievement(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_VictoryPanelAddCustomStatisticLine(LPJASS j){ (void)j; return jass_pushnull(j); }
-static DWORD sc2_VictoryPanelAddTrackedStatistic(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_IntLoopBegin(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_IntLoopDone(LPJASS j)   { return jass_pushboolean(j, true); }
-static DWORD sc2_IntLoopEnd(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_IntLoopStep(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PreloadAsset(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PreloadImage(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PreloadModel(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PreloadMovie(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PreloadObject(LPJASS j) { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PreloadScene(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PreloadScript(LPJASS j) { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PreloadSound(LPJASS j)  { (void)j; return jass_pushnull(j); }
-
-/* -------------------------------------------------------------------------
- * Named no-op stubs (one per unimplemented native for grep-ability)
- * ------------------------------------------------------------------------- */
-static DWORD sc2_AchievementAward(LPJASS j)              { (void)j; return jass_pushnull(j); }
-static DWORD sc2_AchievementErase(LPJASS j)              { (void)j; return jass_pushnull(j); }
-static DWORD sc2_AchievementPanelSetCategory(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_AchievementPanelSetVisible(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_AchievementPercentText(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_AchievementTermQuantitySet(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_AchievementsDisable(LPJASS j)           { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ActorCreate(LPJASS j)                   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ActorFrom(LPJASS j)                     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ActorFromScope(LPJASS j)                { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ActorRegionCreate(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ActorRegionSend(LPJASS j)               { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ActorScopeFromUnit(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ActorSend(LPJASS j)                     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_AIDisableAllScouting(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_BankExists(LPJASS j)                    { (void)j; return jass_pushboolean(j, false); }
-static DWORD sc2_BankKeyRemove(LPJASS j)                 { (void)j; return jass_pushnull(j); }
-static DWORD sc2_BankLastCreated(LPJASS j)               { (void)j; return jass_pushnullhandle(j, "bank"); }
-static DWORD sc2_BankLoad(LPJASS j)                      { (void)j; return jass_pushnullhandle(j, "bank"); }
-static DWORD sc2_BankSave(LPJASS j)                      { (void)j; return jass_pushnull(j); }
-static DWORD sc2_BankValueSetFromFlag(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_BankValueSetFromInt(LPJASS j)           { (void)j; return jass_pushnull(j); }
-static DWORD sc2_BankValueSetFromString(LPJASS j)        { (void)j; return jass_pushnull(j); }
-static DWORD sc2_BankValueSetFromText(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_CampaignMode(LPJASS j)                  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_CinematicDataRun(LPJASS j)              { (void)j; return jass_pushnull(j); }
-static DWORD sc2_CinematicDataStop(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ConversationDataResetNodeState(LPJASS j)  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ConversationDataResetStateValues(LPJASS j){ (void)j; return jass_pushnull(j); }
-static DWORD sc2_ConversationDataSaveNodeState(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ConversationDataSaveStateValues(LPJASS j) { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ConversationDataStateFixedValue(LPJASS j) { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ConversationDataStateGetValue(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ConversationDataStateIndex(LPJASS j)      { (void)j; return jass_pushinteger(j, 0); }
-static DWORD sc2_ConversationDataStateIndexCount(LPJASS j) { (void)j; return jass_pushinteger(j, 0); }
-static DWORD sc2_ConversationDataStateName(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ConversationDataStateSetValue(LPJASS j)   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ConversationDataStateText(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_DataTableSetString(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_DataTableValueExists(LPJASS j)          { (void)j; return jass_pushboolean(j, false); }
-static DWORD sc2_DialogControlSetPropertyAsText(LPJASS j){ (void)j; return jass_pushnull(j); }
-static DWORD sc2_DialogControlSetVisible(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_FixedToInt(LPJASS j)                    { (void)j; return jass_pushinteger(j, 0); }
-static DWORD sc2_FixedToString(LPJASS j)                 { (void)j; return jass_pushnull(j); }
-static DWORD sc2_FormatNumber(LPJASS j)                  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_GameCheatAllow(LPJASS j)                { (void)j; return jass_pushnull(j); }
-static DWORD sc2_GameGetSpeedValue(LPJASS j)             { (void)j; return jass_pushnumber(j, 1.0f); }
-static DWORD sc2_GameIsDebugOptionSet(LPJASS j)          { (void)j; return jass_pushboolean(j, false); }
-static DWORD sc2_GameIsTestMap(LPJASS j)                 { (void)j; return jass_pushboolean(j, false); }
-static DWORD sc2_GameIsTransitionMap(LPJASS j)           { (void)j; return jass_pushboolean(j, false); }
-static DWORD sc2_GameMapIsBlizzard(LPJASS j)             { (void)j; return jass_pushboolean(j, false); }
-static DWORD sc2_GamePauseAllCharges(LPJASS j)           { (void)j; return jass_pushnull(j); }
-static DWORD sc2_GameSetSeedLocked(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_GameSetSpeedLocked(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_GameSetSpeedValue(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_HelpPanelAddTutorial(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_HelpPanelShowTechTreeRace(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_IntToFixed(LPJASS j)                    { (void)j; return jass_pushnumber(j, 0.0f); }
-static DWORD sc2_IntToString(LPJASS j)                   { (void)j; return jass_pushnull(j); }
-static DWORD sc2_ObjectiveGetPrimary(LPJASS j)           { (void)j; return jass_pushnullhandle(j, "objective"); }
-static DWORD sc2_PingCreate(LPJASS j)                    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_PlayerType(LPJASS j)                    { (void)j; return jass_pushinteger(j, 0); }
-static DWORD sc2_PointSetFacing(LPJASS j)                { (void)j; return jass_pushnull(j); }
-static DWORD sc2_SoundChannelSetVolume(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_StringReplaceWord(LPJASS j)             { (void)j; return jass_pushnull(j); }
-static DWORD sc2_StringSub(LPJASS j)                     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_StringToText(LPJASS j)                  { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TechTreeAbilityAllow(LPJASS j)          { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TechTreeAbilityIsAllowed(LPJASS j)      { (void)j; return jass_pushboolean(j, true); }
-static DWORD sc2_TechTreeRestrictionsEnable(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TechTreeUnitHelp(LPJASS j)              { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TechTreeUnitHelpDefault(LPJASS j)       { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TechTreeUpgradeCount(LPJASS j)          { (void)j; return jass_pushinteger(j, 0); }
-static DWORD sc2_TextCase(LPJASS j)                      { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TimerPause(LPJASS j)                    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TransmissionSetOption(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_TriggerDebugOutput(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UIClearMessages(LPJASS j)               { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UIFlyerHelperClearOverride(LPJASS j)    { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UIFlyerHelperOverride(LPJASS j)         { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UIFrameVisible(LPJASS j)                { (void)j; return jass_pushboolean(j, false); }
-static DWORD sc2_UISetCursorVisible(LPJASS j)            { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UISetGameMenuItemVisible(LPJASS j)      { (void)j; return jass_pushnull(j); }
-static DWORD sc2_UISetRestartLoadingScreen(LPJASS j)     { (void)j; return jass_pushnull(j); }
-static DWORD sc2_VisEnable(LPJASS j)                     { (void)j; return jass_pushnull(j); }
+#include "galaxy_math.h"
+#include "galaxy_trigger.h"
+#include "galaxy_camera.h"
+#include "galaxy_cinematic.h"
+#include "galaxy_unit.h"
+#include "galaxy_point.h"
+#include "galaxy_player.h"
+#include "galaxy_sound.h"
+#include "galaxy_transmission.h"
+#include "galaxy_catalog.h"
+#include "galaxy_game.h"
+#include "galaxy_ui.h"
+#include "galaxy_techtree.h"
+#include "galaxy_actor.h"
 
 /* -------------------------------------------------------------------------
  * Native table
